# perf(all): use strings.Cut to replace strings.SplitN